implement generic messagepack-rpc endpoint
Some messagepack-rpc servers, like Neovim, require the client to be able
to handle requests and notifications. This was not possible with the
previous implementation: an endpoint was either a client or a server.

This commit implements generic endpoints that can act both like a client
or a server. It is still possible to create "normal" client and servers.