do not call Task::notify() ourself!
ref:
https://users.rust-lang.org/t/my-future-eats-all-the-cpu-how-can-i-prevent-this/13084/7

The call to notify() was added to "fix" the "ping_pong" example, that
was hanging. I thought that I had to explicitely tell the executor to
re-poll the Endpoint. While that fixes the hanging issue, it makes the
Endpoint being constantly polled which eats all the CPU.

The thing is, when using the Tokio event loop, everything is handled for
us: the Endpoint is polled only when some IO event allows it to makes
some progress. Calling notify() ourself is defeating the purpose of the
Tokio event loop.

This commit also fixes a bug @Fylwind found while helping me debugging
the ping_pong example: we have to read all the messages from the stream
before returning Async::NotReady, otherwise, the Endpoint won't be
polled before a new message arrives, hence leaving these messages unread
for an undetermined period of time.