endpoint: poll streams until NotReady
A stream needs to be polled in a loop until it can no longer make
progress. Only when it returns Async::NotReady has it registered its
event source with the main loop.

The effect of not looping was that only one notification would send per
call to poll(), and poll() would never be re-triggered due to pending
notifications after sending one. Other activity on the socket might
cause poll() to run and allow progress, but without that it would hang.

Signed-off-by: Tyler Hall <tylerwhall@gmail.com>