endpoint: gracefully drop the connection on error
We should not kill the server if there is an error on a single stream.

One case this can happen is if the client sets SO_LINGER to 0 before
closing its socket. The client will send a RST packet and the server
will get "Connection reset by peer" rather than a normal end of stream.

Signed-off-by: Tyler Hall <tylerwhall@gmail.com>