tweak Service::new() signature
Do not pass the tokio::reactor::Handle() by reference.

This does not change much because anyway the service needs to own the
handle and ends up cloning it. It also makes it to write small helper
function that create endpoint like:

> type SessionFuture<H> =
>     Box<Future<Item = Endpoint<H, TcpStream>, Error = ()>>;
>
> pub fn new_tcp<H>(addr: &SocketAddr, handler: H, handle: Handle)
>     -> SessionFuture<H>
> where
>     H: ServiceWithClient + 'static,
> {
>     let endpoint = TcpStream::connect(&addr, &handle)
>         .map(|stream| Endpoint::new(stream, handler, handle))
>         .map_err(|_| ());
>     Box::new(endpoint)
> }

If we just pass a reference do Endpoint::new, rustc complains that it
does not live for the 'static lifetime.